tests/wpt/web-platform-tests/css/css-box/parsing/margin-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS basic box model: margin sets longhands</title>
<link rel="help" href="https://drafts.csswg.org/css-box-3/#propdef-margin">
<meta name="assert" content="margin supports the full grammar '<length-percentage>{1,4}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
</head>
<body>
<script>
test_shorthand_value('margin', '1px 2px 3px 4px', {
  'margin-top': '1px',
  'margin-right': '2px',
  'margin-bottom': '3px',
  'margin-left': '4px'
});

test_shorthand_value('margin', '1px 2px 3px', {
  'margin-top': '1px',
  'margin-right': '2px',
  'margin-bottom': '3px',
  'margin-left': '2px'
});

test_shorthand_value('margin', '1px 2px', {
  'margin-top': '1px',
  'margin-right': '2px',
  'margin-bottom': '1px',
  'margin-left': '2px'
});

test_shorthand_value('margin', '1px', {
  'margin-top': '1px',
  'margin-right': '1px',
  'margin-bottom': '1px',
  'margin-left': '1px'
});
</script>
</body>
</html>
